# backend/docs/swagger.yaml
basePath: /admin-api
definitions:
  github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      msg:
        example: ""
        type: string
    type: object
...
info:
  contact: {}
  description: 'Go modular-monolith scaffold for operations platforms. Default database:
    MySQL 8.4.'
  title: Nimbus Framework Go API
  version: "1.0"
paths:
...
  /member/address/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List a member's addresses
      tags:
      - Member Detail
  /member/experience-record/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Get a member experience record
      tags:
      - Member Detail
  /member/experience-record/page:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Page member experience records
      tags:
      - Member Detail
...
  /member/point/record/page:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Page member point records
      tags:
      - Member Detail
  /member/sign-in/record/page:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Page member sign-in records
      tags:
      - Member Detail
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"